backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
crewai==1.10.0
langchain==0.1.20
langchain-openai==0.1.8
python-dotenv==1.0.0
gitpython==3.1.40
PyGithub==1.59.1
pyflakes==3.0.1
pylint==3.0.3
autopep8==2.0.4
black==23.12.1
isort==5.13.2
pytest==7.4.3
pytest-json-report==1.5.0
httpx==0.25.2
pydantic==2.5.3
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
openai==1.3.0